+++ news-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Contao news bundle change log
 
+### DEV
+
+ * Always show the "show from" and "show until" fields (see contao/core-bundle#908).
+
 ### 4.4.0-RC2 (2017-06-12)
 
  * Trigger all the callbacks in the toggleVisibility() methods (see contao/core-bundle#756).